Refactor story data normalization into dedicated module
